adds exercise updating and workout generating engines
This commit adds all of the functionality for exercise updating and working generating, as well as additional testing for these features, both using Jest and Postman. This also incorporates these features into the /current_workout endpoint, which will return the current workout for the user if it exists and generate it if it doesn't.